Add multy message

+++ src/main/java/life/good/goodlife/model/map/Location.java
@@ -1,22 +1,22 @@
 package life.good.goodlife.model.map;
 
 public class Location {
-    private double lat;
-    private double lng;
+    private float lat;
+    private float lng;
 
-    public double getLat() {
+    public float getLat() {
         return lat;
     }
 
-    public void setLat(double lat) {
+    public void setLat(float lat) {
         this.lat = lat;
     }
 
-    public double getLng() {
+    public float getLng() {
         return lng;
     }
 
-    public void setLng(double lng) {
+    public void setLng(float lng) {
         this.lng = lng;
     }
 }
